renamed admin controllers, format with prettier
